Creating new SMV utils
